move repeated code in redis and valkey backends into a base class
